requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-flat-responsive==1.4.1
django-rest-framework==0.1.0
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.3
pytz==2017.2
whitenoise==3.3.0
